requirements.txt
flask
flask-admin
flask-assets
flask-debugtoolbar
flask-markdown
flask-script
flask-security
flask-sqlalchemy
flask-wtf
